fix: Script name changed to __main__.py.
Added __init__.py but it is empty. The script now can read the route config/queries.yml and it gets stored into a var called config
