+++ app/requirements.txt
@@ -1,2 +1,7 @@
+certifi==2025.8.3
+charset-normalizer==3.4.3
 fpdf==1.7.2
+idna==3.10
 requests==2.32.4
+urllib3==2.5.0
+pyyaml==6.0.2
